# contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_enrich_artifacts__artifact_id__enrich_patch": {
        "properties": {
          "training_table": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Table"
          }
        },
        "title": "Body_enrich_artifacts__artifact_id__enrich_patch",
        "type": "object"
      },
      "Body_ingest_artifacts_ingest_post": {
        "properties": {
          "eval_labels": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eval Labels"
          },
          "eval_samples": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eval Samples"
          },
          "labels": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels"
          },
          "model": {
            "contentMediaType": "application/octet-stream",
            "description": "LightGBM .txt model file",
            "title": "Model",
            "type": "string"
          },
          "model_name": {
            "description": "Logical model name, e.g. 'loan_default_classifier'",
            "title": "Model Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "samples": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Samples"
          },
          "training_table": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Table"
          }
        },
        "required": [
          "model",
          "model_name"
        ],
        "title": "Body_ingest_artifacts_ingest_post",
        "type": "object"
      },
...
      "EnrichResponse": {
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "enriched_at": {
            "title": "Enriched At",
            "type": "string"
          },
          "sections_added": {
            "items": {
              "type": "string"
            },
            "title": "Sections Added",
            "type": "array"
          }
        },
        "required": [
          "artifact_id",
          "sections_added",
          "enriched_at"
        ],
        "title": "EnrichResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestResponse": {
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "sections": {
            "$ref": "#/components/schemas/SectionStatus"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "model_id",
          "model_name",
          "version",
          "artifact_id",
          "sections"
        ],
        "title": "IngestResponse",
        "type": "object"
      },
...
      "ModelDetail": {
        "description": "Detail page data \u2014 full version history, newest first.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionSummary"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "versions",
          "created_at"
        ],
        "title": "ModelDetail",
        "type": "object"
      },
      "ModelSummary": {
        "description": "Card data for the registry home view.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "framework": {
            "title": "Framework",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latest_version": {
            "title": "Latest Version",
            "type": "integer"
          },
          "latest_version_date": {
            "title": "Latest Version Date",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "objective": {
            "title": "Objective",
            "type": "string"
          },
          "section_status": {
            "$ref": "#/components/schemas/SectionStatus"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "latest_version",
          "latest_version_date",
          "framework",
          "objective",
          "section_status",
          "created_at"
        ],
        "title": "ModelSummary",
        "type": "object"
      },
...
      "SectionStatus": {
        "properties": {
          "data_profile": {
            "title": "Data Profile",
            "type": "boolean"
          },
          "evaluation": {
            "title": "Evaluation",
            "type": "boolean"
          },
          "importance": {
            "title": "Importance",
            "type": "boolean"
          },
          "shap": {
            "title": "Shap",
            "type": "boolean"
          },
          "trees": {
            "title": "Trees",
            "type": "boolean"
          }
        },
        "required": [
          "trees",
          "importance",
          "shap",
          "evaluation",
          "data_profile"
        ],
        "title": "SectionStatus",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionSummary": {
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "section_status": {
            "$ref": "#/components/schemas/SectionStatus"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "artifact_id",
          "section_status",
          "notes",
          "created_at"
        ],
        "title": "VersionSummary",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Model introspection platform \u2014 canonical artifact API. Every response cites the schema version it conforms to.",
    "title": "Cerebro",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/artifacts/ingest": {
      "post": {
        "operationId": "ingest_artifacts_ingest_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_artifacts_ingest_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest"
      }
    },
...
    "/artifacts/{artifact_id}/enrich": {
      "patch": {
        "description": "Add a data profile section to an existing artifact.\n\nSHAP and evaluation enrichment require re-ingesting the model file via\nPOST /artifacts/ingest \u2014 the API layer cannot import LightGBM (invariant #2).",
        "operationId": "enrich_artifacts__artifact_id__enrich_patch",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_enrich_artifacts__artifact_id__enrich_patch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrichResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enrich"
      }
    },
...
    "/models": {
      "get": {
        "operationId": "list_models_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "framework",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Framework"
            }
          },
          {
            "in": "query",
            "name": "objective",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Objective"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelSummary"
                  },
                  "title": "Response List Models Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Models",
        "tags": [
          "models"
        ]
      }
    },
    "/models/{model_id}": {
      "get": {
        "operationId": "get_model_models__model_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model",
        "tags": [
          "models"
        ]
      }
    },
    "/models/{model_id}/versions": {
      "get": {
        "operationId": "list_versions_models__model_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VersionSummary"
                  },
                  "title": "Response List Versions Models  Model Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Versions",
        "tags": [
          "models"
        ]
      }
    }
  }
}